% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_plate_replicates}
\alias{add_plate_replicates}
\title{Replicate reactions across multiple plates}
\usage{
add_plate_replicates(reactions, plate_replicate_number_total)
}
\arguments{
\item{reactions}{A data frame containing reactions to be replicated across plates.}

\item{plate_replicate_number_total}{The total number of plate replicates.}
}
\value{
A data frame with reactions replicated across the specified number of plates,
        including a new identifier for each plate replicate.
}
\description{
This function takes a data frame of reactions and adds replicates for each reaction
based on the total number of plates. It adds an identifier for plate replicates.
}
\examples{
reactions_df <- data.frame(ReactionID = 1:10)
replicated_df <- add_plate_replicates(reactions_df, 3)

}
