% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_plates}
\alias{add_plates}
\title{Renumber wells across multiple plates}
\usage{
add_plates(df, well_index_column, plate_size)
}
\arguments{
\item{df}{Data frame containing plate well indexes.}

\item{well_index_column}{Name of the column with well indexes (as an unquoted string).}

\item{plate_size}{Number of wells per plate.}
}
\value{
Data frame with added `Plate.Index` indicating plate number
        and renumbered `Well.Index` within each plate.
}
\description{
Renumber wells across multiple plates
}
\examples{
reaction_data <- data.frame(Well.Index = 1:96)
updated_data <- addPlates(reaction_data, "Well.Index", 96)
}
