% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPlates.R
\name{addPlates}
\alias{addPlates}
\title{Plot Plates}
\usage{
addPlates(dat, Well.Index.ColumnID, Plate.Size)
}
\arguments{
\item{dat}{A data frame containing the well index}

\item{Well.Index.ColumnID}{The column name of well index}

\item{Plate.Size}{The size of plate for this set of wells}
}
\value{
An dataframe where with an additional column Plate.Index and Well.Index <= Plate.Size
}
\description{
Plot Plates
}
