% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_well.R
\name{convert_well}
\alias{convert_well}
\title{Convert between well labels and numeric indices}
\usage{
convert_well(input, direction, plate = 96, wise = "row")
}
\arguments{
\item{input}{A vector of well labels (as character) or numeric indices (as numeric) to be converted.}

\item{direction}{Character; specifies the conversion direction, either "to_num" or "to_well".}

\item{plate}{The size of the well plate (96, 384, or 1536 are supported).}

\item{wise}{Character; specifies direction the plate is filled ("row" or "col" wise).
Applicable for both "to_num" and "to_well" directions.}
}
\value{
A vector of converted well labels or numeric indices.
}
\description{
Convert between well labels and numeric indices
}
