% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_designer.R
\name{design_plate}
\alias{design_plate}
\title{Interactive Shiny app to design plate layouts}
\usage{
design_plate()
}
\value{
A [shiny::shinyApp()] object.
}
\description{
Launches a Shiny application for designing microtiter plate layouts: pick a
plate type, number of reactions, replicates and spacing, and get back a
destination plate map as a plot and a downloadable table.
}
\section{Status}{

This GUI is **experimental**. The programmatic functions it wraps
([replicate_reactions()], [add_plates()], [add_plate_replicates()],
[convert_well()], [plot_plate()]) are the supported interface.
}

