% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_designer.R
\name{plate_designer_server}
\alias{plate_designer_server}
\title{Plate Design Server}
\usage{
plate_designer_server(
  id,
  createPlateMap,
  rxnCount,
  replicatecount,
  destplate,
  deststartwell,
  wellspacing,
  replicatestyle,
  fillwise,
  platereplicates
)
}
\arguments{
\item{id}{Unique identifier for the server module.}

\item{createPlateMap}{Reactive expression linking the 'Create Plate Map' button.}

\item{rxnCount}{Reactive value for the number of reactions.}

\item{replicatecount}{Reactive value for the number of replicates per reaction.}

\item{destplate}{Reactive value for the type of destination plate.}

\item{deststartwell}{Reactive value for the starting well of the plate.}

\item{wellspacing}{Reactive value for the space between wells on the plate.}

\item{replicatestyle}{Reactive value for the prioritization strategy of replicates.}

\item{fillwise}{Reactive value indicating row-wise or column-wise filling.}

\item{platereplicates}{Reactive value for the number of plate replicates.}
}
\value{
Shiny module that encapsulates the side-effects and rendering of plot and datatables based on inputs.
}
\description{
The server function for the plate designer Shiny module. It processes inputs
from the UI and updates the output of the destination plate design. It generates
visual or tabular representations of the designed plate.
}
