% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_designer.R
\name{plate_designer_ui}
\alias{plate_designer_ui}
\title{Plate Design UI Module}
\usage{
plate_designer_ui(id, label = "Setup")
}
\arguments{
\item{id}{Unique identifier for UI elements to facilitate modularization.}

\item{label}{Label for the setup section, default is "Setup".}
}
\description{
This UI module creates a user interface for the plate design application. It
provides input fields and selectors for configuring the design parameters of
a destination microtiter plate including type of plate, number of reactions and
replicates, and well spacing.
}
