% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlateSizeCheck.R
\name{platesize.check}
\alias{platesize.check}
\title{Check the entered plate size is allowable}
\usage{
platesize.check(plate)
}
\arguments{
\item{plate}{Numeric, the number of wells in the plate}
}
\value{
a list containing the row ids, column ids and max number for both rows and columns
}
\description{
Check the entered plate size is allowable
}
