% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/platesize_check.R
\name{platesize_check}
\alias{platesize_check}
\title{Check allowable plate size and return dimensions}
\usage{
platesize_check(plate_size)
}
\arguments{
\item{plate_size}{Numeric, the number of wells in the plate.}
}
\value{
A list containing the information of plate dimensions:
          - rows: A character vector of row identifiers.
          - columns: A numeric vector of column numbers.
          - rowmax: The total number of rows in the plate.
          - colmax: The total number of columns in the plate.
}
\description{
This function validates if the entered plate size is supported and returns the
corresponding rows and columns dimensions.
}
