% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_plate.R
\name{plot_plate}
\alias{plot_plate}
\title{Plot Plates}
\usage{
plot_plate(
  data,
  fill,
  well_id,
  facet_rows = NULL,
  facet_cols = NULL,
  plate = 384,
  size = 5,
  shape = 22,
  na_fill = "white",
  na_size_ratio = 0.95,
  na_alpha = 0.1
)
}
\arguments{
\item{data}{A data frame containing the well locations and the values to fill}

\item{fill}{The column name of the value to fill consider wrapping in double brackets and quotes}

\item{well_id}{The column name of the character column of well locations}

\item{facet_rows}{The column name of the character column selected for the row facet}

\item{facet_cols}{The column name of the character column selected for the column facet}

\item{plate}{The size of the plate}

\item{size}{The size of the plotted data on the plate}

\item{shape}{The shape of each plotted data point on the plate}

\item{na_fill}{Fill color for the plate}

\item{na_size_ratio}{Background ratio}

\item{na_alpha}{Background transparency}
}
\value{
An image object
}
\description{
Plot Plates
}
