% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotPlates.R
\name{plotPlate}
\alias{plotPlate}
\title{Plot Plate}
\usage{
plotPlate(
  Data,
  fill = fill,
  wellID = wellID,
  plate = 384,
  size = 5,
  shape = 22,
  na_fill = "white",
  na_size_ratio = 0.95,
  na_alpha = 0.1
)
}
\arguments{
\item{Data}{A data frame containing the well locations and the values to fill}

\item{fill}{The column name of the value to fill consider wrapping in double brackets and quotes}

\item{wellID}{The column name of the character column of well locations}

\item{plate}{The size of the plate}

\item{size}{The size of the plotted data on the plate}

\item{shape}{The shape of each plotted data point on the plate}

\item{na_fill}{Fill color for the plate}

\item{na_size_ratio}{Background ratio}

\item{na_alpha}{Background transparency}
}
\value{
An image object
}
\description{
Plot Plate
}
