% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_reaction_positions.R
\name{replicate_reactions}
\alias{replicate_reactions}
\title{Replicate Reactions and Assign Unique Numeric Positions}
\usage{
replicate_reactions(
  rxn_df,
  reaction_col_name,
  num_replicates,
  priority,
  inter_spacing,
  intra_spacing,
  start_position
)
}
\arguments{
\item{num_replicates}{The number of times each reaction should be replicated.}

\item{priority}{The priority for positioning: "reaction" for reaction-first ordering,
or "replicate" for replicate-first ordering.}

\item{inter_spacing}{The numeric distance between the sets, either different reactions or replicates.}

\item{intra_spacing}{The numeric distance between replicates within a set.}

\item{start_position}{The starting numeric position for the first entry.}

\item{reactions_vector}{A vector of unique reaction identifiers to be replicated.}
}
\value{
A data frame with the columns 'reaction', 'replicate', and 'position'.
        Each row corresponds to a reaction replicate with a unique assigned position.
}
\description{
This function takes a vector of unique reaction identifiers and a number of replicates, and
assigns a unique numeric position to each reaction based on the specified ordering priority.
}
\examples{
data <- data.frame(rxn = 1:3)
data <- replicate_reactions(data, "rxn", num_replicates = 3, priority = "reaction",
                            inter_spacing = 10, intra_spacing = 1, start_position = 1)
data <- replicate_reactions(data, "rxn", num_replicates = 3, priority = "replicate",
                            inter_spacing = 5, intra_spacing = 1, start_position = 1)

}
