% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toNum.R
\name{to.num}
\alias{to.num}
\title{to.num}
\usage{
to.num(well, wise, plate = 96)
}
\arguments{
\item{well}{A character vector of well locations to be converted to numbers}

\item{wise}{"row" or "col" - Indicates the direction the plate is read, across (L->R) a row or down a column}

\item{plate}{The size of the well plate (96 and 384 are supported)}
}
\value{
A numeric vector of the well locations on the plate
}
\description{
to.num
}
