% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toWell.R
\name{to.well}
\alias{to.well}
\title{To Well}
\usage{
to.well(num, wise, plate = 96)
}
\arguments{
\item{num}{Numeric - a number less than the plate size representing the location on the well plate}

\item{wise}{"row" or "col" - Indicates the direction the plate is read, across (L->R) a row or down a column}

\item{plate}{The size of the well plate (96 and 384 are supported)}
}
\value{
A character vector representing the row location with a letter and the column location with a number
}
\description{
To Well
}
