% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_rows_columns}
\alias{to_rows_columns}
\title{Convert Well IDs to Row and Column Indices}
\usage{
to_rows_columns(wells, well_id)
}
\arguments{
\item{wells}{A dataframe containing a column with well IDs.}

\item{well_id}{A string representing the name of the column in `wells` with the well IDs.}
}
\value{
A tibble with two additional columns `Row` and `Column` for the numeric row and column indices.
}
\description{
This function takes a dataframe and a column name containing well IDs (e.g., "A1", "B6", "AA1") and
returns the dataframe with two new columns for the row and column indices based on those well IDs.
}
\examples{
wells_df <- data.frame(WellID = c("A1", "B6", "AA10", "AZ25"))
wells_df <- to_rows_columns(wells_df, "WellID")

}
