% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPlate.R
\name{toRowsColumns}
\alias{toRowsColumns}
\title{Convert to Row and Column indices}
\usage{
toRowsColumns(Wells, wellID)
}
\arguments{
\item{Wells}{The dataframe containing the wells to be converted}

\item{wellID}{The ColName of the column containing the character values to be converted}
}
\value{
Two columns added to the dataframe for the row and column indices
}
\description{
Convert to Row and Column indices
}
