% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plate_designer.R
\name{design_plate}
\alias{design_plate}
\title{R Shiny App to Design Plate Formats}
\usage{
design_plate()
}
\description{
This function creates a Shiny application for designing microtiter plate formats.
The application allows users to set up various parameters for plate design and
generates a destination plate map as output.
}
